fix: keep the task completion validator out of the renderer daemon import chain
The renderer api-client imported gui-result-validation, which transitively
loads daemon modules and the preset package; the preset resolves an assets
URL from import.meta.url at module load and throws under the jsdom GUI test
runner (21 renderer test files red in CI). Move validateDaemonTaskCompletion
into daemon-protocol-gui-types next to the other GUI payload validators and
point the schema registry and the renderer at that module.
